Refactor: File moving, split Goban into render base and light Goban interface for engine callbacks
