+++ jest.config.ts
@@ -31,7 +31,7 @@ export default {
     // ],
     coveragePathIgnorePatterns: [
         "/src/test.tsx",
-        "/src/goban/index.ts",
+        "/src/renderer/index.ts",
         "/src/engine/index.ts",
         ".d.ts",
         "wasm_estimator.ts",
@@ -82,9 +82,7 @@ export default {
     // maxWorkers: "50%",
 
     // An array of directory names to be searched recursively up from the requiring module's location
-    // moduleDirectories: [
-    //   "node_modules"
-    // ],
+    moduleDirectories: ["src", "src/third_party", "src/third_party/goscorer", "node_modules"],
 
     // An array of file extensions your modules use
     // moduleFileExtensions: [
