+++ package.json
@@ -25,11 +25,11 @@
         "dev": "webpack-cli serve",
         "build-debug": "webpack",
         "build-production": "webpack --mode production",
-        "dts": "dts-bundle-generator -o lib/goban.d.ts  src/goban/index.ts",
+        "dts": "dts-bundle-generator -o lib/goban.d.ts  src/index.ts",
         "lint": "eslint src/ --ext=.ts,.tsx",
         "lint:fix": "eslint --fix src/ --ext=.ts,.tsx",
-        "typedoc": "typedoc src/goban.ts",
-        "typedoc:watch": "typedoc --watch src/goban.ts",
+        "typedoc": "typedoc src/index.ts",
+        "typedoc:watch": "typedoc --watch src/index.ts",
         "prettier": "prettier --write \"src/**/*.{ts,tsx}\"",
         "prettier:check": "prettier --check \"src/**/*.{ts,tsx}\"",
         "checks": "npm run lint && npm run prettier:check",
