+++ src/__tests__/GoConditionalMove.test.ts
@@ -1,4 +1,4 @@
-import { GoConditionalMove } from "../GoConditionalMove";
+import { GoConditionalMove } from "../engine";
 
 /**
  * ```
