+++ src/__tests__/GoEngine_sgf.test.ts
@@ -11,8 +11,8 @@
 
 (global as any).CLIENT = true;
 
-import { TestGoban } from "../TestGoban";
-import { MoveTree } from "../MoveTree";
+import { TestGoban } from "../renderer/TestGoban";
+import { MoveTree } from "../engine";
 
 type SGFTestcase = {
     template: string;
