+++ src/engine/ScoreEstimator.ts
@@ -18,9 +18,9 @@ import { encodeMove } from "./GoMath";
 import * as GoMath from "./GoMath";
 import { StoneString } from "./StoneString";
 import { StoneStringBuilder } from "./StoneStringBuilder";
-import type { GobanCore } from "../goban/GobanCore";
-import { GoEngine, PlayerScore, GoEngineRules } from "./GoEngine";
-import { JGOFMove, JGOFNumericPlayerColor, JGOFSealingIntersection } from "./JGOF";
+import type { Goban } from "../Goban";
+import { GoEngine, PlayerScore, GoEngineRules } from "./GobanEngine";
+import { JGOFMove, JGOFNumericPlayerColor, JGOFSealingIntersection } from "./formats/JGOF";
 import { _ } from "./translate";
 import { wasm_estimate_ownership, remote_estimate_ownership } from "./ownership_estimators";
 import * as goscorer from "goscorer";
@@ -128,7 +128,7 @@ export class ScoreEstimator extends BoardState {
 
     constructor(
         engine: GoEngine,
-        goban_callback: GobanCore | undefined,
+        goban_callback: Goban | undefined,
         trials: number,
         tolerance: number,
         prefer_remote: boolean = false,
