+++ src/renderer/themes/board_plain.ts
@@ -16,7 +16,7 @@
 
 import { GoTheme, GoThemeBackgroundCSS } from "../GoTheme";
 import { GoThemesInterface } from "../GoThemes";
-import { callbacks } from "engine/callbacks";
+import { callbacks } from "../callbacks";
 import { _ } from "engine/translate";
 
 // Converts a six-digit hex string to rgba() notation
