+++ src/renderer/themes/image_stones.ts
@@ -20,7 +20,7 @@ import { _ } from "engine/translate";
 import { deviceCanvasScalingRatio, allocateCanvasOrError } from "../canvas_utils";
 import { renderShadow } from "./rendered_stones";
 import { renderPlainStone } from "./plain_stones";
-import { callbacks } from "engine/callbacks";
+import { callbacks } from "../callbacks";
 
 const anime_black_imagedata = makeSvgImageData(require("../../../assets/img/anime_black.svg"));
 const anime_white_imagedata = makeSvgImageData(require("../../../assets/img/anime_white.svg"));
