+++ src/renderer/themes/plain_stones.ts
