+++ src/renderer/themes/rendered_stones.ts
