+++ test/test_autoscore.ts
@@ -29,9 +29,13 @@
 import { existsSync, readFileSync, readdirSync } from "fs";
 import { autoscore } from "../src/engine/autoscore";
 import * as clc from "cli-color";
-import { GoEngine, GoEngineInitialState } from "../src/engine/GoEngine";
+import { GoEngine, GoEngineInitialState } from "../src/engine/GobanEngine";
 import { char2num, makeMatrix, num2char } from "../src/engine/GoMath";
-import { JGOFMove, JGOFNumericPlayerColor, JGOFSealingIntersection } from "../src/engine/JGOF";
+import {
+    JGOFMove,
+    JGOFNumericPlayerColor,
+    JGOFSealingIntersection,
+} from "../src/engine/formats/JGOF";
 
 function run_autoscore_tests() {
     const test_file_directory = "autoscore_test_files";
