+++ tsconfig.node.json
@@ -15,6 +15,7 @@
         "baseUrl": ".",
         "paths": {
             "*": ["src/*", "*"],
+            "engine": ["src/engine"],
             "goscorer": ["src/third_party/goscorer/goscorer"]
         },
         "noImplicitAny": true,
